+++ src/examples/click_counter/main.ts
@@ -1,0 +1,7 @@
+// main.ts
+
+import { Clicker } from "./clicker.ts";
+import { PreactUsage } from "./preact_usage.ts";
+
+document.body.appendChild(Clicker());
+document.body.appendChild(PreactUsage());