+++ src/index.ts
@@ -1,4 +1,3 @@
 // src/index.ts
 
 export * from "./core";
-export * from "./devtools";